cek session login admin

+++ application/controllers/Barang.php
@@ -12,17 +12,24 @@ class Barang extends CI_Controller {
 
 	public function index(){
 
-		$data['barang'] = $this->Model_barang->get();
-		
-		$this->load->view('admin/head');
-		$this->load->view('barang/index', $data);
-		$this->load->view('datatables');
-		$this->load->view('javascript-admin');
+		if ($this->session->userdata('admin') == 1) {
+
+			$data['barang'] = $this->Model_barang->get();
+
+			$this->load->view('admin/head');
+			$this->load->view('barang/index', $data);
+			$this->load->view('datatables');
+			$this->load->view('javascript-admin');
+		}else {
+			$data['heading'] = "<center><h1>404</h1></center>";
+			$data['message'] = "<center><p>Page Not Found</p></center>";
+			$this->load->view('errors/html/error_404', $data);
+		}
 	}
 
 	public function barang_json_get(){
 		$data = $this->Model_barang->get();
-// mengambil data dari json, (cocok untuk web service)
+		// mengambil data dari json, (cocok untuk web service)
 		$this->output
 		->set_content_type('application/json')
 		->set_output(json_encode($data))
@@ -58,50 +65,56 @@ class Barang extends CI_Controller {
 	}
 
 	public function edit($id){
+		if ($this->session->userdata('admin') == 1) {
 
-		if ($this->input->server('REQUEST_METHOD') == 'POST') {
-			$data['nama_barang']	= $this->input->post('nama_barang');
-			$data['id_kategori']	= $this->input->post('id_kategori');
-			$data['harga']			= $this->input->post('harga');
-			$data['keterangan']		= $this->input->post('keterangan');
+			if ($this->input->server('REQUEST_METHOD') == 'POST') {
+				$data['nama_barang']	= $this->input->post('nama_barang');
+				$data['id_kategori']	= $this->input->post('id_kategori');
+				$data['harga']			= $this->input->post('harga');
+				$data['keterangan']		= $this->input->post('keterangan');
 
-			if(!$_FILES['foto']['name'] == '')
-			{
-          //upload file config
-				$config['upload_path'] = 'products';
-				$config['allowed_types'] = 'jpg|png';
-				$config['max_size'] = 5000;
-				$config['encrypt_name'] = TRUE;
-
-				$this->load->library('upload', $config);
-
-          //uploading File
-				if(!$this->upload->do_upload('foto'))
+				if(!$_FILES['foto']['name'] == '')
 				{
-					$this->session->set_flashdata('error', $this->upload->display_errors());
-					redirect('admin/barang/' . $id, 'refresh');
-				}
-				else
-				{
-            // delete image File
-					$path = "products/";
-					$record = $this->Model_barang->select_byid($id);
-					$filename = $record[0]->foto;
-					unlink($path . $filename);
+					//upload file config
+					$config['upload_path'] = 'products';
+					$config['allowed_types'] = 'jpg|png';
+					$config['max_size'] = 5000;
+					$config['encrypt_name'] = TRUE;
 
-					$data['foto'] = $this->upload->data()['file_name'];
+					$this->load->library('upload', $config);
+
+					//uploading File
+					if(!$this->upload->do_upload('foto'))
+					{
+						$this->session->set_flashdata('error', $this->upload->display_errors());
+						redirect('admin/barang/' . $id, 'refresh');
+					}
+					else
+					{
+						// delete image File
+						$path = "products/";
+						$record = $this->Model_barang->select_byid($id);
+						$filename = $record[0]->foto;
+						unlink($path . $filename);
+
+						$data['foto'] = $this->upload->data()['file_name'];
+					}
 				}
+
+				$this->Model_barang->edit($data, $id);
+				redirect('admin/barang','refresh');
 			}
 
-			$this->Model_barang->edit($data, $id);
-			redirect('admin/barang','refresh');
+			$data['barang'] = $this->Model_barang->select_byid($id);
+
+			$this->load->view('admin/head');
+			$this->load->view('barang/barang_edit', $data);
+
+		}else {
+			$data['heading'] = "<center><h1>404</h1></center>";
+			$data['message'] = "<center><p>Page Not Found</p></center>";
+			$this->load->view('errors/html/error_404', $data);
 		}
-
-		$data['barang'] = $this->Model_barang->select_byid($id);
-
-		$this->load->view('admin/head');
-		$this->load->view('barang/barang_edit', $data);
-
 	}
 
 	public function hapus($id){
